requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pytesseract==0.3.10
pdf2image==1.16.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
xmltodict==0.13.0
streamlit==1.28.0